requirements.txt
azure-cognitiveservices-speech
python-dotenv
sounddevice
numpy
pyaudio
scipy
wavio
spacy
datapackage